Define sources and sinks
Sources and sinks are an experimental alternative to transformers.
If we define NxSink (not done yet), a transformer's loading can be
defined as a source feeding into a NxSink.  If we also define NxSource,
we can define a transformer's saving as a NxSource feeding into some
target format sink.

This decoupling should enable a streaming model that requires less
memory to transform large datasets.